Rename `TableAssets` -> `TableExplorer`
