+++ web-common/src/layout/navigation/Navigation.svelte
@@ -13,7 +13,7 @@
   import { writable } from "svelte/store";
   import AddAssetButton from "../../features/entity-management/AddAssetButton.svelte";
   import FileExplorer from "../../features/file-explorer/FileExplorer.svelte";
-  import TableAssets from "../../features/tables/TableAssets.svelte";
+  import TableExplorer from "../../features/tables/TableExplorer.svelte";
   import Resizer from "../Resizer.svelte";
   import { DEFAULT_NAV_WIDTH } from "../config";
   import Footer from "./Footer.svelte";
@@ -66,7 +66,7 @@
         <FileExplorer />
         <div class="grow" />
         {#if navWrapperHeight}
-          <TableAssets startingHeight={navWrapperHeight / 2} />
+          <TableExplorer startingHeight={navWrapperHeight / 2} />
         {/if}
       </div>
     </div>
